refactor: standardize entry points and project structure - Move main logic to root index.js - Standardize types to types/index.d.ts - Update internal test imports to match new structure - Update package.json entry points
