+++ package.json
@@ -2,9 +2,9 @@
   "name": "fastify-plugin",
   "version": "5.1.0",
   "description": "Plugin helper for Fastify",
-  "main": "plugin.js",
+  "main": "index.js",
   "type": "commonjs",
-  "types": "types/plugin.d.ts",
+  "types": "types/index.d.ts",
   "scripts": {
     "lint": "eslint",
     "lint:fix": "eslint --fix",
